add voi_manager
